[Django Admin] Move admin-only views from "admin_views.py" to the DailyChallenge's ModelAdmin
