+++ src/apps/daily_challenge/urls.py
@@ -1,6 +1,6 @@
 from django.urls import path
 
-from . import admin_views, views
+from . import views
 
 app_name = "daily_challenge"
 
@@ -37,17 +37,4 @@ urlpatterns = [
     # Debug views (staff only)
     path("debug/reset-today", views.debug_reset_today),
     path("debug/view-cookie", views.debug_view_cookie),
-    # Admin views (staff only)
-    # TODO: use ModelAdmin.get_urls instead?
-    #   https://docs.djangoproject.com/en/4.2/ref/contrib/admin/#django.contrib.admin.ModelAdmin.get_urls
-    path(
-        "admin/daily-challenge/game-preview",
-        admin_views.preview_daily_challenge,
-        name="admin_game_preview",
-    ),
-    path(
-        "admin/daily-challenge/play-future-daily-challenge/<str:lookup_key>/",
-        admin_views.play_future_daily_challenge,
-        name="play_future_daily_challenge",
-    ),
 ]
